change password

+++ app/controllers/settingsCtrl.js
@@ -8,6 +8,7 @@
         $scope.selectedCategories;
         $scope.categoryList = [];
         $scope.insuranceDetails={};
+        $scope.passwordDetails={};
         $scope.cardDetailIndicator = false;
         $scope.updateCardDetailIndicator = false;
         $scope.buttonIndicator = true;
@@ -315,27 +316,22 @@
         });
     }
 
-    $scope.updatePassword = function (resetForm) {
-       // var userId = commonService.getUserid();
-//        httpService.updateProfile(userId, $scope.userDetails).then(function (response) {
-//            if (response.data.message == "Profile updated successfully") {
-//                $scope.successPasswordIndicator = true;
-//            }
-//        });
-        resetForm.$setSubmitted(true);
-        if (resetForm.$valid) {
+    $scope.updatePassword = function (passwordDetails,passForm) {
+
+        passForm.$setSubmitted(true);
+        if (passForm.$valid) {
               console.log("here");
             $scope.passwordIndicator = false;
-         
-            if ($scope.cnfNewPassword === $scope.newPassword) {
+          console.log('scope',$scope.passwordDetails);
+            if ($scope.passwordDetails.cnfNewPassword === $scope.passwordDetails.newPassword) {
                 $rootScope.loaderIndicator = true;
                 var userId = commonService.getUserid();
-                httpService.changePassword(userId, $scope.oldPassword, $scope.newPassword).then(function (result) {
-                    if (result.data.code != 500 || result.data.message == "Password changed successfully.") {
+                httpService.changePassword(userId, $scope.passwordDetails.oldPassword, $scope.passwordDetails.newPassword).then(function (result) {
+                    if (result.data.code != 500 && result.data.message == "Password changed successfully.") {
                         $rootScope.loaderIndicator = false;
-                        $scope.oldPassword = "";
-                        $scope.newPassword = "";
-                        $scope.cnfNewPassword = "";
+                        $scope.passwordDetails.oldPassword = "";
+                        $scope.passwordDetails.newPassword = "";
+                        $scope.passwordDetails.cnfNewPassword = "";
                         $scope.successIndicator = true;
                         commonService.deleteCookieValues('FirstName');
                         commonService.deleteCookieValues('UserID');
@@ -343,10 +339,10 @@
                         $rootScope.isLogin = false;
                         $rootScope.adminIndicator = false;
                         commonService.reloadRoute();
-                       // $location.path('/');
-                       // $('#promptLoginPopup').modal('toggle');
-                      //  $("#promptLoginPopup").modal({ backdrop: "static" });
-                      //  $('#promptLoginPopup').modal('show');
+                        $location.path('/');
+                        $('#promptLoginPopup').modal('toggle');
+                        $("#promptLoginPopup").modal({ backdrop: "static" });
+                        $('#promptLoginPopup').modal('show');
                     }
                     else {
                         $rootScope.loaderIndicator = false;
